Replaced manual allocation by regions.
Wrapper functions around glGen/Create* used to call malloc to
capture a pointer as long as the object is not out of scope. That
is a waste of space, because we don’t need to keep track of the
pointer. We only need the GLuint value.

Replaced malloc/free with alloca/with.