Added attachmentLength into FramebufferAttachment.
That function is used to determine the length of an attachment type.

For color attachments, it’ll recursively scan down the tuple of formats
(:.) and gets its length.

For depth attachments, because depth attachment should be singleton, the
size is either 0 or 1.

We can use that function to re-implement hasColor and hasDepth by
comparing to 0 (>0).