Made the `U` interface safer.
The idea was to remove the Uniformed type as it was way too much
permissive on pure effects. Now, the `U` safety relies on
contravariance and divisibility.

Because `U` is a (divisible and decidable) contravariant functor, we
can embed new types into `U` by dividing them and contramapping. We
then end up with a large `U` type we can directly updates on the
GPU, which is awesome!